feat(nextjs): Add support for mdx pages
The index.tsx page is now index.mdx. This required a change to the babel config as the JSX pragma in use caused issues with the mdx loader but this issue is being worked on. https://github.com/mdx-js/mdx/pull/1413

We seem to now require React to be explicitly imported